pyproject.toml
[tool.poetry]
name = "gist"
version = "0.1.0"
description = "Summarize and allow queries on Web pages."
authors = ["Gist Author <author@example.com>"]
readme = "README.md"
packages = [{ include = "gist", from = "apps" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
django = ">=5.2.5,<6.0.0"
requests = ">=2.32.5,<3.0.0"
beautifulsoup4 = ">=4.13.4,<5.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
gunicorn = ">=22.0.0,<23.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.12.9"
ipython = "^9.4.0"
pytest-django = "^4.11.1"
pytest = "^8.4.1"
pytest-mock = "^3.14.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.7"

[tool.black]
target-version = ['py312']
exclude = '''
/(\
    \.git
    | \.mypy_cache
    | \.venv
    | build
    | dist
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test_*.py"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"

exclude = [
	"__pycache__",
	"migrations",
	"db.sqlite3",
	"*.pyc"
]

[tool.ruff.lint]
select = ["E", "F", "I"]
# E501 (line too long) is ignored
ignore = ["E501"]
